+++ CHANGELOG.md
@@ -10,6 +10,7 @@ script to proceed.
 - Improved error logging.
 - Move `temboard-migratedb` as `temboard migratedb` subcommand.
 - Move `temboard-agent-register` as `temboard-agent register` subcommand.
+- New command `temboard generate-key` for generating signing key.
 - New subcommands `temboard query-agent`, `temboard routes`, `temboard
   runtask`, `temboard schedule`, `temboard web`, `temboard-agent routes`,
   `temboard-agent runtask` and `temboard-agent web` for debugging.